Hero Image bug fix
Currently he hero image thrashes the ui on inital loading when site is deployed. This is most likely due to absolute positioning of animated elements but, I'm not sure. So using Svelte-Image to set a placeholder while loading to avoid this.

WIP - currently working just need to fix figure styling around image and image sizing.